fix: core set brightness and timestamp were checking unproperly device capabilities and return Core error -5: No error, now it's working properly

+++ lib/core/drivers/driver.hpp
@@ -89,7 +89,7 @@ public:
      */
     virtual bool setTemperature(
         const std::string& uuid,
-        uint32_t value
+        float value
     ) = 0;
 
     /**
